[JBMETA-390]   Moved validation classes to common project   Enables each project to provide it's own json file for schema   validation   Cleanup, new packages etc
