+++ common/src/test/java/org/jboss/test/metadata/common/vdx/ValidationHelper.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.metadata.ejb.test.common;
+package org.jboss.test.metadata.common.vdx;
 
 import org.w3c.dom.Document;
 import org.xml.sax.InputSource;
